linux/32: Add a define to use llseek instead of _llseek
Some architecutres (such as RV32) use the name "llseek" for the 62nd
llseek syscall isntead of "_llseek". This can be checked by looking at
the output in the generated ksysent.h file and seeing this:

  #ifdef __NR_llseek
  [__NR_llseek & 0xffff] = "llseek",
  #endif

Add a new define that defaults to being set to 1 that allows us to
toggle between "_llseek" and "llseek". RV32 will then set the define to
0.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>